requirements.txt
allure-pytest
requests~=2.27.1
pytest~=6.2.5
simplejson~=3.17.6